userInit.m
Added the checkpoint mechanism in order not to check all the points when moving the robot. The idea is when moving the robot, it will go on straight line until the next checkpoint. Once it reaches the checkpoint, the robot stops and turns until the heading points to the next checkpoint.